app/requirements.txt
Flask>=3.1,<4.0
groq>=1.2.0,<2.0.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0
PyJWT>=2.8,<3.0
